[AN-2059] move data object from inside the datalist to the same level as the datalist in the getUserTimeline response.
