+++ dataprovider/src/main/java/cm/aptoide/pt/dataprovider/model/v7/TimelineData.java
@@ -1,0 +1,20 @@
+package cm.aptoide.pt.dataprovider.model.v7;
+
+/**
+ * Created by jdandrade on 28/12/2017.
+ */
+
+public class TimelineData {
+  private String version;
+
+  public TimelineData() {
+  }
+
+  public String getVersion() {
+    return version;
+  }
+
+  public void setVersion(String version) {
+    this.version = version;
+  }
+}
